app/requirements.txt
flask
requests
aerospike
celery[redis]
redis
python-dotenv